+++ src/components/LinksPanel/components/ChatDrawer/ChatDrawer.test.tsx
@@ -1,5 +1,5 @@
 import { act, fireEvent, render, renderHook, screen, waitFor } from '@testing-library/react';
-import { createSelector, getJestSelectors } from '@volkovlabs/jest-selectors';
+import { createSelector, getJestSelectors } from '@/test-utils/jest-selectors';
 import React from 'react';
 
 import { TEST_IDS } from '@/constants';