+++ src/components/LinksPanel/components/LinksLayout/LinksLayout.test.tsx
@@ -1,6 +1,6 @@
 import { FieldType, toDataFrame } from '@grafana/data';
 import { act, render, screen } from '@testing-library/react';
-import { createSelector, getJestSelectors } from '@volkovlabs/jest-selectors';
+import { createSelector, getJestSelectors } from '@/test-utils/jest-selectors';
 import React from 'react';
 
 import { TEST_IDS } from '@/constants';