+++ src/components/editors/GroupsEditor/components/GroupEditor/components/LinkEditor/components/TimePickerWrapper/TimePickerWrapper.test.tsx
@@ -1,5 +1,5 @@
 import { act, render, screen } from '@testing-library/react';
-import { createSelector, getJestSelectors } from '@volkovlabs/jest-selectors';
+import { createSelector, getJestSelectors } from '@/test-utils/jest-selectors';
 import React from 'react';
 import { config } from '@grafana/runtime';
 import { TEST_IDS } from '@/constants';